Refactor code to allow users to upgrade their subscriptions
This commit refactors the code in the `librewoo-upgrade.php` file to allow users to upgrade their subscriptions. It introduces a new class `LibreSignUpgrade` with a constructor that adds an action hook to the `woocommerce_add_to_cart` event. The `upgrade_subscription` method checks if the user is logged in and retrieves the current user ID and cart items. It then checks if the user has any active subscriptions for the same product and allows the purchase if the cart product ID is greater than the subscription product ID. The commit also includes a `custom_add_to_cart_button_text` method that changes the add to cart button text to "Upgrade".

The changes in this commit enable users to easily upgrade their subscriptions and provide a better user experience.